==> 1. Data preparation.Rmd <==
---
output:
  html_notebook: default
  html_document: default
---
# Install packages


```{r packages}
library(data.table)
library(dplyr)
library(ggplot2)
library(psych) # where 'describe' comes from
library(lme4)
library(usethis)
library(lmerTest)
library(nlme)
library(MuMIn)
library(car)
library(MASS)
```



# data collection

## average national milk price in the Netherlands 
Source : https://ec.europa.eu/info/food-farming-fisheries/farming/facts-and-figures/markets/overviews/market-observatories/milk_en
Regulation (EU) No 2017/1185 Article 12(a) - Annex II.4(a)

The annual average price was calculated from monthly price by me.

```{r milk price}
TestYear <- c(2005:2016)
MilkPricePer100KgInMarket <- c(28.69,27.81,32.82,35.92,26.77,31.46,36.88,33.80,39.45,39.44,30.60,28.38)
MilkPriceInMarket <- data.frame(TestYear,MilkPricePer100KgInMarket)
```



## import raw data where accounting data, herd data and herd lactatation curve charateristics have already merged together 
```{r data import}
RawData <- fread('./data/AccountingWithUBNWithProductionWithSCCWithCIWithLCC.csv')
SoilType <- fread('./data/SoilType.csv')

#add SoilType to the whole dataset
RawData <- merge(RawData, SoilType, by = c('KVK','FiscalYear'), all.x=T)

#add OfficalMilkPrice to the whole dataset
RawData <- merge(RawData, MilkPriceInMarket, by = c('TestYear'), all.x=T)
```


```{r}
colnames(RawData)
```

# data calculation

```{r calculation}
CalculatedData <- RawData %>% 
    group_by(UBN) %>% 
    filter(TotalMilkForFactoryKg != 0) %>% 
    mutate(FeedCostPerCow = TotalFeedCosts / FlynthNumberOfMilkingCow,
          MilkRevenuePerCow = MilkRevenue / FlynthNumberOfMilkingCow,
          IncomeOverFeedCostPerCow = MilkRevenuePerCow-FeedCostPerCow,
          MilkPerFTE = TotalMilkInFarmKg/TotalFTE,
          HerdIntensity = TotalMilkInFarmKg/TotalLandAreaInHa,
          OutsourcingYoungStockRearing = ifelse(RearingCostsOfYoungStockForThirdParties != 0, "Yes", "No"),
          SoilType = ifelse(SoilType %in% 1, 'SandSoil', ifelse(SoilType %in% 2:4, 'OtherSoil', NA)),
          RelativeMilkPrice = (100 * MilkRevenue / TotalMilkForFactoryKg ) - MilkPricePer100KgInMarket,
          GrossMarginPer100kg = 100 * GrossMargin / TotalMilkForFactoryKg,
          LastyearInFlynth = lag(FiscalYear, 1 , order_by = FiscalYear),
          NextyearInFlynth = lead(FiscalYear, 1 , order_by = FiscalYear),
          ExpansionRatePerYear = ExpansionRate / (FiscalYear - LastyearInFlynth),
          ) %>%
    ungroup()
    
```


```{r}
#Fill NA expansion rate with mean of the year or 0 (2007 or non-2007)

CalculatedData <- CalculatedData %>%
    group_by(FiscalYear) %>%
    mutate_at(vars(ExpansionRate, ExpansionRatePerYear),~ifelse(FiscalYear==2007, 0, 
                                                                ifelse(is.na(.x), mean(.x, na.rm = TRUE), .x))) %>%
    ungroup()
   
```


# save data

```{r}
save(CalculatedData, file="CalculatedData.rda")
```
